+++ src/codegen/generator.rs
@@ -266,7 +266,7 @@ pub trait Generator {
 		enum_type_id: TypeId,
 		enum_shape_index: usize,
 		enum_specialization_index: usize,
-		tag_value: u64,
+		tag_value: i128,
 		variant_binding: Option<Self::Binding>,
 	) -> Self::Binding;
